created a new module for frequency thresholding
